### 2018-07-16 #22 @whedon
[PRE REVIEW]: Quantitative Big Imaging
**Submitting author:** @kmader (<a href="http://orcid.org/0000-0003-4919-9359">Kevin Mader</a>)
**Repository:** <a href="https://github.com/kmader/Quantitative-Big-Imaging-2018/tree/jose-submission" target ="_blank">https://github.com/kmader/Quantitative-Big-Imaging-2018/tree/jose-submission</a>
**Version:** v2018
