### 2018-07-16 #22 @labarba
@kmader : I believe you've got the metadata section of the paper wrong. See this published paper, as an example:
https://raw.githubusercontent.com/arm61/pylj/master/paper/paper.md
