### 2018-07-16 #22 @kmader
hmm, I tried to copy as closely as possible: https://raw.githubusercontent.com/kmader/Quantitative-Big-Imaging-2018/jose-submission/paper.md
I can't seem to figure out from the ruby messages what is going on, is it because it isn't the master branch?
