### 2018-07-16 #22 @labarba
I actually don't know the answer to your question. I know that the directory location does not matter, as `whedon` looks for the `paper.md` file wherever it may be. But I don't know if it works on a branch! That's a question for @arfon, but he's on vacation until 22 July.
