### 2018-07-16 #22 @labarba
On the other hand, it doesn't seem like a good idea to give us a URL to a branch in the "Submission repository," as this carries over to the Review issue and than to the published paper's page on the journal. 😬
