### 2018-07-16 #22 @kyleniemeyer
I do think the paper needs to be in master, though I'm not 100% sure.
