### 2018-07-16 #22 @kmader
Can I change it or do I need to delete and resubmit? I was hoping to keep the JOSE part off the master branch until it was finalized but it's not a problem.
