### 2018-07-16 #22 @labarba
If you merge it to master, I can manually edit this issue and try to recompile, to see if it works!
