### 2018-07-16 #22 @kyleniemeyer
You can just change it, then prompt a new attempt to compile with `@whedon generate pdf`
