### 2018-07-24 #22 @arfon
> I do think the paper needs to be in master, though I'm not 100% sure.

👍 the paper needs to be in the default Git branch which for the vast majority of repositories is `master`.
