### 2018-07-31 #22 @katyhuff
Hi folks. I can handle editing this but it will be a few days before I will be able to give it my full attention. Thank you in advance for your patience.
