### 2018-08-07 #22 @katyhuff
@arokem and @ThomasA: I think you would both be excellent reviewers for [this JOSE submission](https://github.com/openjournals/jose-papers/blob/jose.00022/jose.00022/10.21105.jose.00022.pdf) regarding Quantitative Big Imaging.

The Journal of Open Source Education Reviewer Guidelines can be found [here](http://jose.theoj.org/about#reviewer_guidelines), and are very similar to the guidelines for JOSS. Please note that JOSE accepts two types of submissions: (1) computational learning modules, created as open educational resources (OER), and (2) open-source software, created as educational technology or infrastructure. This one appears to be a submission of the first type.

What do you think: would you be willing to review this submission to JOSE?
