### 2018-08-08 #25 @whedon
[REVIEW]: Quantitative Big Imaging
**Submitting author:** @kmader (<a href="http://orcid.org/0000-0003-4919-9359">Kevin Mader</a>)
**Repository:** <a href="https://github.com/kmader/Quantitative-Big-Imaging-2018" target ="_blank">https://github.com/kmader/Quantitative-Big-Imaging-2018</a>
**Version:** v2018

- [x] **Version:** Does the release version given match the repository release (v2018)?

- [ ] **Version:** Does the release version given match the repository release (v2018)?
