### 2018-08-08 #25 @kmader
@labarba Not that I know of, I don't immediately see what the error message is, is something I can change or try?
