### 2018-08-08 #25 @labarba
I think the message means that the `paper` file is not being found. @arfon ?
