### 2018-08-08 #25 @arfon
It's because this URL doesn't resolve:

https://github.com/kmader/Quantitative-Big-Imaging-2018/tree/jose-submission
