### 2018-08-08 #25 @arfon
I've updated the URL of the repository address to https://github.com/kmader/Quantitative-Big-Imaging-2018 which has fixed this.
