### 2018-08-13 294ce83
Create travis_install.sh
adding the first bit of travis code
