+++ circleci/travis_install.sh
@@ -1,0 +1,37 @@
+# Adapted shamelessly from https://github.com/scikit-learn-contrib/project-template/blob/master/ci_scripts/install.sh
+# Deactivate the travis-provided virtual environment and setup a
+# conda-based environment instead
+deactivate
+
+# Use the miniconda installer for faster download / install of conda
+# itself
+pushd .
+cd
+mkdir -p download
+cd download
+echo "Cached in $HOME/download :"
+ls -l
+echo
+if [[ ! -f miniconda.sh ]]
+   then
+   wget http://repo.continuum.io/miniconda/Miniconda3-latest-Linux-x86_64.sh \
+       -O miniconda.sh
+   fi
+chmod +x miniconda.sh && ./miniconda.sh -b
+cd ..
+export PATH="$HOME/miniconda3/bin:$PATH"
+conda update --yes conda
+popd
+
+# Configure the conda environment and put it in the path using the
+# provided versions
+conda create -n testenv --yes python=$PYTHON_VERSION pip nose pytest
+conda env update -n testenv -f binder/environment.yml -q
+source activate testenv
+
+if [[ "$COVERAGE" == "true" ]]; then
+    pip install pytest-cov coverage coveralls
+fi
+chmod a+x binder/postBuild
+sh binder/postBuild
+© 2018 GitH
