### 2018-08-13 d13b095
Update travis_install.sh
remove pytest code
