### 2018-08-13 0b2e67d
Update .travis.yml
forcing kernel for older or misformatted notebooks
