+++ .travis.yml
@@ -16,7 +16,7 @@ install: source circleci/travis_install.sh
 script:
   - shopt -s globstar
   - for nb in **/*ipynb; do
-    jupyter nbconvert --ExecutePreprocessor.timeout=3600 --execute "$nb" --to markdown |& tee nb_to_md.txt;
+    jupyter nbconvert --ExecutePreprocessor.timeout=3600 --ExecutePreprocessor.kernel_name=testkernel --execute "$nb" --to markdown |& tee nb_to_md.txt;
     traceback=$(grep "Traceback (most recent call last):" nb_to_md.txt);
     if [[ $traceback ]]; then
         exit 1;
