### 2018-08-13 beecdb7
Update README.md
adding travis build tag
