### 2018-08-14 f6f4f0a
Update travis_install.sh
removing paste error
