### 2018-08-14 0c5134f
updating to circle 20
first draft of config file
