### 2018-08-14 cc3f21f
Delete circle.yml
