+++ circle.yml
@@ -1,17 +1,0 @@
-dependencies:
-  override:
-    - bash circleci/install.sh:
-        timeout: 1024
-
-test:
-  # Run and convert examples to markdown, then grep for errors
-  override:
-    - bash circleci/execute.sh:
-        timeout: 3600
-    - if grep -q "Traceback (most recent call last):" nb_to_md.txt; then false; else true; fi
-
-
-general:
-  # Open the doc to the API
-  artifacts:
-    - ~/notebooks
