### 2018-08-14 c24a314
Update .travis.yml
fixing kernel setup
