### 2018-08-14 7ac1b6e
Update .travis.yml
skipping files in kaggle folder
