### 2018-09-23 kmader/Quantitative-Big-Imaging-2018#19 @arokem
[JOSE review]: Statement of need
The section in the paper titled "statement of need" doesn't quite state the need for these materials. Maybe explain what kinds of gaps the course from which these materials were derived filled in the curriculum that this was part of?
