### 2018-09-23 kmader/Quantitative-Big-Imaging-2018#20 @arokem
[JOSE Review]: Target audience
I couldn't find a clear statement of the target audience
