### 2018-09-23 kmader/Quantitative-Big-Imaging-2018#21 @arokem
[JOSE review]: Learning objectives
The learning objectives are not clearly delineated. What specific tasks would the students being taught with these materials be able to do at the end of each lesson?
