### 2018-09-23 kmader/Quantitative-Big-Imaging-2018#22 @arokem
[JOSE review]: List of software dependencies
Though there are local references to dependencies in various parts of the materials (including in slides and in videos) It would be good to have a list of all the software dependencies in one place (e.g., in the README file).
