### 2018-09-23 kmader/Quantitative-Big-Imaging-2018#23 @arokem
[JOSE review]: Image enhancement app is a bit wonky
The binder link from lecture 2 sends me to a binder instance, that shows what looks like a widget that then disappears, and a help menu appears. I think that something might be wrong with that thing.
