### 2018-09-23 #25 @arokem
This submission includes a set of materials that were created for the teaching of an image processing course at the ETH in Spring 2016. The course includes lectures about a large range of topics in image processing, introducing modern methods in computational science along the way. From what I could gather from watching the lecture videos and going through the materials, it looks like the course was fantastic. I wish that these topics would have been taught to me early on in this way.

As I said, the materials provided in the repository, including lecture videos, notebooks, and slide-decks are excellent, and could potentially be made very useful as a set of reusable materials that others could pick and choose from, when designing courses, or even as materials for self-teaching. However, as the submission is currently organized, it carries with it too much of the legacy of its original provenance as it was written to accompany a particular course in a particular time and place. I think that for this to be a valuable contribution in this context, and to facilitate reusability, some of this specific context might need to be set aside, in favor of a more general set of materials, targetted at reuse.

For example, I think that many of the slide-decks are not currently useful to anyone who wasn't in the class, without additional notes that would tell us what to do with these slides. For example, the slides for the second lecture are useful as an accompaniment to the lecture that Dr. Kaestner gave, but are not necessarily useful to anyone without detailed notes from that lecture. A similar issue arises with Dr. Prummer's slides (from May 17th), but even worse, because no video is provided for this lecture. Similar can be said for many other lecture materials, that do not stand on their own, and I believe cannot currently be effectively reused, without substantial additional effort.

An overall guide in the readme that tells potential users of these materials (learners and instructors) how they should use the materials would be useful. Detailed narrative notes to accompany the lectures or at least more notes in the slides and notebooks would help.

The profusion of materials and lack of guidance on how to proceed through the course (either as a learner or as an instructor) is further exacerbated by the presence of deprecated materials from previous years. I would recommend to remove these materials, or at least relegate them to somewhere less prominent.

Because materials are spread between Youtube, Github, Kaggle and other places I may have missed, I worry also that some of these materials may not be available in the long run. This dispersion of the materials would also make it impossible to archive the entire set of materials together in something like a Zenodo DOI.

I would really love to see these materials better organized because I think that this set of materials could potentially serve as a very valuable contribution to JOSE, but I think that a major revision of the materials would be required to achieve that goal.
