### 2018-09-23 #25 @arokem
See also:
https://github.com/kmader/Quantitative-Big-Imaging-2018/issues/19
https://github.com/kmader/Quantitative-Big-Imaging-2018/issues/20
https://github.com/kmader/Quantitative-Big-Imaging-2018/issues/21
https://github.com/kmader/Quantitative-Big-Imaging-2018/issues/22
https://github.com/kmader/Quantitative-Big-Imaging-2018/issues/23
