### 2018-11-12 #25 @labarba
Hi  @kmader — 👋
Wondering here what your sense is on re-organizing the materials for better re-usability, according to @arokem's review. Thanks!
