### 2018-11-14 #25 @kmader
@labarba @arokem thanks for the feedback, I definitely agree with all of it and appreciate the effort you have put in. The changes look fairly substantial and I am currently swamped with other projects. I was hoping to get to some of the issues over the coming holiday breaks.
