### 2018-12-10 c65aa10
Adding more clear dependencies list
Adding a complete list of software dependencies for reproducibility.
- closes #22
