### 2018-12-10 c1ffe2f
Create full_package_list.txt
a frozen version of the full package list
