### 2018-12-10 16ab482
Update README.md
moving requirements to wiki page
