### 2018-12-10 kmader/Quantitative-Big-Imaging-2018#22 @kmader
I have added a more detailed list of software dependencies including an exact list of all packages and versions to the [wiki page](https://github.com/kmader/Quantitative-Big-Imaging-2018/wiki/Full-Package-List) and made a [Docker Hub image](https://hub.docker.com/r/kmader/qbi2018/) from the latest repo to embody any dependencies that might somehow disappear in the future
