### 2018-12-10 6c8e43c
Locking sklearn 0.19
locking older version of sklearn (0.20 breaks predict on single values)
