+++ binder/environment.yml
@@ -22,7 +22,7 @@ dependencies:
   - pydot
   - h5py
   - scikit-image<=0.20
-  - scikit-learn<=0.20
+  - scikit-learn==0.19
   - pytorch<=0.3.0
   - torchvision
   - opencv3>=3.0
