### 2018-12-10 5c7fb61
Delete full_package_list.txt
moved to wiki
