+++ binder/full_package_list.txt
@@ -1,178 +1,0 @@
-# packages in environment at /Users/kevinmader/miniconda3/envs/qbi2018:
-#
-# Name                    Version                   Build  Channel
-absl-py                   0.1.10                     py_0    conda-forge
-appmode                   0.3.0                     <pip>
-appnope                   0.1.0                    py36_0    conda-forge
-asn1crypto                0.24.0                   py36_0    conda-forge
-astor                     0.6.2                      py_0    conda-forge
-autopep8                  1.3.5                     <pip>
-awscli                    1.15.57                   <pip>
-backcall                  0.1.0                      py_0    conda-forge
-beautifulsoup4            4.6.3                     <pip>
-blas                      1.1                    openblas    conda-forge
-bleach                    1.5.0                    py36_0    conda-forge
-bokeh                     0.12.16                  py36_0    conda-forge
-botocore                  1.10.56                   <pip>
-bs4                       0.0.1                     <pip>
-ca-certificates           2018.4.16                     0    conda-forge
-certifi                   2018.4.16                py36_0    conda-forge
-cffi                      1.11.5                   py36_0    conda-forge
-chardet                   3.0.4                    py36_0    conda-forge
-clangdev                  6.0.0                 default_0    conda-forge
-click                     6.7                       <pip>
-cloudpickle               0.5.3                     <pip>
-colorama                  0.3.9                     <pip>
-cryptography              2.2.1                    py36_0    conda-forge
-cycler                    0.10.0                   py36_0    conda-forge
-dask-core                 0.17.5                     py_0    conda-forge
-dask-ndfilters            0.1.2                     <pip>
-dask-ndmeasure            0.1.1                     <pip>
-dask-ndmorph              0.1.1                     <pip>
-deap                      1.2.2                     <pip>
-decorator                 4.3.0                      py_0    conda-forge
-distributed               1.21.8                    <pip>
-docutils                  0.14                      <pip>
-entrypoints               0.2.3                    py36_1    conda-forge
-ffmpeg                    2.8.6                         0    menpo
-freetype                  2.8.1                         0    conda-forge
-gast                      0.2.0                      py_0    conda-forge
-graphviz                  2.38.0                        7    conda-forge
-grpcio                    1.11.0           py36hd9629dc_0  
-h5py                      2.7.1                    py36_1    conda-forge
-hdf5                      1.8.18                        3    conda-forge
-HeapDict                  1.0.0                     <pip>
-html5lib                  0.9999999                py36_0    conda-forge
-icu                       58.2                          0    conda-forge
-idna                      2.6                      py36_1    conda-forge
-imageio                   2.3.0                    py36_0    conda-forge
-ipykernel                 4.8.2                    py36_0    conda-forge
-ipython                   6.4.0                    py36_0    conda-forge
-ipython_genutils          0.2.0                    py36_0    conda-forge
-ipyvolume                 0.4.5                    py36_0    conda-forge
-ipywebrtc                 0.3.0                    py36_0    conda-forge
-ipywidgets                7.2.1                    py36_1    conda-forge
-itk                       4.13.0                   py36_0    conda-forge
-itkwidgets                0.9.0                     <pip>
-jedi                      0.12.0                   py36_0    conda-forge
-jinja2                    2.10                     py36_0    conda-forge
-jmespath                  0.9.3                     <pip>
-jpeg                      9b                            2    conda-forge
-jsonschema                2.6.0                    py36_1    conda-forge
-jupyter_client            5.2.3                    py36_0    conda-forge
-jupyter_contrib_core      0.3.3                    py36_1    conda-forge
-jupyter_contrib_nbextensions 0.5.0                    py36_0    conda-forge
[…]
