### 2018-12-10 6176d73
Update README.md
adding target audience #20 and learning objectives #21. Added a very brief statement of need #19
