### 2018-12-10 9cdbdf6
fixing wonkiness of image enhancement playground
closes #23
