### 2019-01-08 112a1b8
Update README.md
adding registration section to readme and fixing a link
