### 2019-02-09 #25 @labarba
Followed up by email with @arokem now.

👋  @ThomasA — I don't see any checked items on your list. Have you been able to work on this review for JOSE? Give us an update when you can. Thanks!
