### 2019-02-09 #25 @arokem
Sorry for the slowness. I will comment on that PR.

Any other thoughts about the comments I made [above](https://github.com/openjournals/jose-reviews/issues/25#issuecomment-423789775)? It looks like much of my feedback in these comments has not been incorporated yet.
