### 2019-02-10 #25 @kmader
@arokem I focused on the concrete issues first because they were easier to address. As for the comments

> For example, I think that many of the slide-decks are not currently useful to anyone who wasn't in the class, without additional notes that would tell us what to do with these slides. For example, the slides for the second lecture are useful as an accompaniment to the lecture that Dr. Kaestner gave, but are not necessarily useful to anyone without detailed notes from that lecture. A similar issue arises with Dr. Prummer's slides (from May 17th), but even worse, because no video is provided for this lecture. Similar can be said for many other lecture materials, that do not stand on their own, and I believe cannot currently be effectively reused, without substantial additional effort.

Unfortunately I was only able to ensure recordings on the lecture I held myself and thus the others were not done.

> An overall guide in the readme that tells potential users of these materials (learners and instructors) how they should use the materials would be useful. Detailed narrative notes to accompany the lectures or at least more notes in the slides and notebooks would help.

Detailed narrative notes would be nice but the time required to remake all of the slides notes and narratives would be on the order of weeks to months (which I don't imagine having in the foreseeable future)

> The profusion of materials and lack of guidance on how to proceed through the course (either as a learner or as an instructor) is further exacerbated by the presence of deprecated materials from previous years. I would recommend to remove these materials, or at least relegate them to somewhere less prominent.

The course is meant to be followed chronologically as most lectures build upon ideas in the previous, but many students manage it out of order. I kept the profusion of materials at the specific request of students who found it challenging to go between 4 years of websites. In addition, given the diversity of students in the class (ranging from computer science grad students to Art History undergrads) means it was always necessary to offer multiple parallel paths of exercises in Python (for more advanced students) and KNIME/ImageJ (for less technical ones). I tried to keep these clearly labeled but it is quite a lot to keep straight.

> Because materials are spread between Youtube, Github, Kaggle and other places I may have missed, I worry also that some of these materials may not be available in the long run. This dispersion of the materials would also make it impossible to archive the entire set of materials together in something like a Zenodo DOI.

Unfortunately it would be exceptionally difficult to have a real, useful course on a single platform. All have their strengths and weaknesses and this certainly makes long run sustainability more challenging, it was done for the benefit of the students in the course today.
- Quantitative 'Big' Imaging requires hosting of datasets and computing resources which exceed what github and mybinder currently offer (thus kaggle datasets and kaggle kernels)
- Reproducibility requires that students are able to setup their own environment without being forced to put their data on a commercial site thus mybinder and repo2docker for the structure so they can recreate the exact setup everywhere and travisci ensures that the setup is valid and every notebook completes without error
- Having videos as piles of MP4 files is quite difficult to store and navigate (42 hours worth) using YouTube provides quick access, playing at 2x speed and subtitles with translations for other languages
