### 2019-02-10 b9ce887
Update README.md
- completely rewriting learning objectives
- adding more details to target audience
