### 2019-02-11 #25 @labarba
Here's the thing, @kmader. You have collected years of materials from your course in one GitHub repository, but this is not what JOSE was designed to publish. One of the key ideas for JOSE publications is that the material be _reusable_ by other instructors or self-learners. Also, we write in the journal's scope:

> **What do you mean by "open-source educational materials"?**
> Examples include Jupyter notebooks or plaintext/markup language documents like LaTeX, R Markdown, and ReST for course/lesson content and associated notes, with embedded or associated code snippets/programs.
> We do not mean openly available slides, lecture notes, or YouTube videos, though these may be acceptable as supplementary materials.

As a general rule, it's nearly impossible for instructors to re-use slide decks prepared by others. We also explicitly exclude multi-media content (e.g., videos), because of the focus on _open source_ content, that can be modified, version controlled, forked and built on.

From your collection of materials, if you were to organize only the Jupyter notebooks into a coherent, self-contained learning module, that's a JOSE publication. Imagine if you were giving an intense, three-day tutorial at a conference, and you polished just the notebooks for that purpose, expecting to do fast-paced  live coding, plus leave the audience with fully narrated materials to review at their own pace after the event. Those materials would be a JOSE module.
