### 2019-02-11 #25 @kmader
Thanks for that info @labarba , I guess in that case it won't make sense to make this repo as it is into a JOSE publication (students still use the repo and it would not be fair to them to remove possibly useful materials to make it conform with JOSE's requirements).
I think I had originally misunderstood the exact purpose of JOSEs work and had incorrectly assumed it was meant to be open-source materials for students (all figures, examples and diagrams driven by tested, continuously integrated code rather than static slides). I had never intended for my course to be particularly useful for lecturers preparing other courses and as such it is probably a bad fit.
When I have time, I'll try to put together some of the material into a new repository just for JOSE, but I guess I can make that as a new separate submission and we can safely reject and close this one.

Thanks in any event for the helpful feedback I think it has made the material more accessible for students.
