### 2019-02-11 #25 @labarba
Here's an idea: you could create a new repository with the Jupyter notebooks, polish them a little bit to make them self-contained with references out to the vids, etc. as supplementary, and then include that repo in your main course one as a GitHub module. The JOSE paper would be associated with the sub-module repo.
