### 2019-03-01 70790cc
Update environment.yml
updating pytorch dep
