+++ binder/environment.yml
@@ -6,6 +6,7 @@ channels:
   - pluraldatum
   - glemaitre
   - soumith
+  - pytorch
   - conda-forge
   - defaults
   - simpleitk
@@ -23,7 +24,7 @@ dependencies:
   - h5py
   - scikit-image<=0.20
   - scikit-learn<=0.20
-  - pytorch<=0.3.0
+  - pytorch<2.0.0
   - torchvision
   - opencv3>=3.0
   - tensorflow
