### 2019-09-28 #25 @labarba
I'm not sure why `whedon` did that. This submission was paused a few months back, as you can read in the thread above. But given the long time lapse, it really makes more sense to close this issue, withdraw the submission. If and when the authors re-organize their materials for a self-contained JOSE submission, they can start a new one.

Thank you for your willingness to review, @ThomasA !
